chore: move `findParentNode` util
